+++ server.js
@@ -2,4 +2,4 @@ const { SocketSignalWebsocketServer } = require('socket-signal-websocket')
 
 // Backward compatibility with v2
 module.exports = (...args) => new SocketSignalWebsocketServer(...args)
-module.exports.SignalSwarmServer = SocketSignalWebsocketServer
+module.exports.SignalServer = SocketSignalWebsocketServer